+++ lib/src/amalgams.dart
@@ -1,27 +1,19 @@
 part of trotter;
 
-/** A pseudo-list of amalgams (permutations with repetition).
- *
- * A pseudo-list "containing" all the [r]-amalgams of objects taken from
- * the list [items].
- * 
- * _Example_
- *
- * ```
- * var a = new Amalgams(3, characters("abcd"));
- * print("There are ${a.length} 3-amalgams of the objects");
- * print("in ${a.items}.");
- * print("The first amalgam is ${a[0]}.");
- * ```
- * 
- */
+/// A pseudo-list of amalgams (permutations with repetition).
+/// 
+/// A pseudo-list "containing" all the [r]-amalgams of objects taken from
+/// the list [items].
+/// 
+/// _Example_
+/// 
+///     var a = new Amalgams(3, characters("abcd"));
+///     print("There are ${a.length} 3-amalgams of the objects");
+///     print("in ${a.items}.");
+///     print("The first amalgam is ${a[0]}.");
+/// 
 
 class Amalgams extends _Combinatorics {
-  int _r;
-
-  /// The number of items taken from [items].
-  int get r => _r;
-
   Amalgams(int r, List items) {
     if (r < 0)
       throw new Exception("Cannot take $r items from ${items.length}.");
@@ -32,6 +24,11 @@ class Amalgams extends _Combinatorics {
     _length = Math.pow(items.length, r).toInt();
   }
 
+  int _r;
+
+  /// The number of items taken from [items].
+  int get r => _r;
+
   @override List operator [](int k) =>
       _amalgam(_adjustedIndex(k, length), r, items);
 
@@ -40,7 +37,8 @@ class Amalgams extends _Combinatorics {
       contains(amalgam) ? _inverseAmalgam(amalgam, _items) : -1;
 
   /// returns whether [x] is in the pseudo-list.
-  bool contains(List x) => _itemsExistInUniversal(x, _items);
+  @override
+  bool contains(Object x) => _itemsExistInUniversal(x, _items);
 
   @override String toString() =>
       "Pseudo-list containing all $length $r-amalgams of items from $items.";