+++ lib/src/compounds.dart
@@ -1,20 +1,17 @@
 part of trotter;
 
-/** A pseudolist of compounds.
- * 
- * A pseudo-list "containing" all the compounds (permutations of 
- * unspecified length) of objects taken from the list `elements`.
- * 
- * _Example_
- *
- * ```
- * var com = new Compounds(letters("abcd"));
- * print("There are ${com.length} subsets of the objects");
- * print("in ${com.items}.");
- * print("The first subset is ${com[0]}.");
- * ```
- * 
- */
+/// A pseudolist of compounds.
+/// 
+/// A pseudo-list "containing" all the compounds (permutations of 
+/// unspecified length) of objects taken from the list `elements`.
+/// 
+/// _Example_
+/// 
+///     var com = new Compounds(letters("abcd"));
+///     print("There are ${com.length} subsets of the objects");
+///     print("in ${com.items}.");
+///     print("The first subset is ${com[0]}.");
+/// 
 
 class Compounds extends _Combinatorics {
   Compounds(List items) {
@@ -35,7 +32,8 @@ class Compounds extends _Combinatorics {
       contains(compound) ? _inverseCompound(compound, _items) : -1;
 
   /// returns whether [x] is in the pseudo-list.
-  bool contains(List x) =>
+  @override
+  bool contains(Object x) =>
       _itemsExistInUniversal(x, _items) && _itemsAreUnique(x);
 
   @override String toString() =>