+++ lib/src/permutations.dart
@@ -1,26 +1,20 @@
 part of trotter;
 
-/** A pseudo-list of permutations.
- *
- * A pseudo-list "containing" all the [r]-permutations of objects taken from
- * the list [items].
- *
- * _Example_
- *
- * ```
- * var p = new Permutations(3, characters("abcd"));
- * print("There are ${p.length} 3-permutations of the objects");
- * print("in ${p.items}.");
- * print("The first permutation is ${p[0]}.");
- * ```
- * 
- */
+/// A pseudo-list of permutations.
+/// 
+/// A pseudo-list "containing" all the [r]-permutations of objects taken from
+/// the list [items].
+/// 
+/// _Example_
+/// 
+///     var p = new Permutations(3, characters("abcd"));
+///     print("There are ${p.length} 3-permutations of the objects");
+///     print("in ${p.items}.");
+///     print("The first permutation is ${p[0]}.");
+/// 
+
 
 class Permutations extends _Combinatorics {
-  int _r;
-
-  /// The number of items taken from [items].
-  int get r => _r;
 
   Permutations(int r, List items) {
     if (r < 0 || r > items.length)
@@ -32,6 +26,11 @@ class Permutations extends _Combinatorics {
     _length = _nPr(items.length, r);
   }
 
+  int _r;
+
+  /// The number of items taken from [items].
+  int get r => _r;
+
   @override List operator [](int k) =>
       _permutation(_adjustedIndex(k, length), r, items);
 
@@ -40,7 +39,8 @@ class Permutations extends _Combinatorics {
       contains(permutation) ? _inversePermutation(permutation, _items) : -1;
 
   /// returns whether [x] is in the pseudo-list.
-  bool contains(List x) =>
+  @override
+  bool contains(Object x) =>
       _itemsExistInUniversal(x, _items) && _itemsAreUnique(x);
 
   @override String toString() =>