+++ lib/src/selections.dart
@@ -1,27 +1,20 @@
 part of trotter;
 
-/** A pseudo-list of selections (combinations with repetition).
- *
- * A pseudo-list "containing" all the [r]-selections of objects taken from
- * the list [items].
- *
- * _Example_
- *
- * ```
- * var s = new Selections(3, characters("abcd"));
- * print("There are ${s.length} 3-selections of the objects");
- * print("in ${s.items}.");
- * print("The first selection is ${c[0]}.");
- * ```
- * 
- */
+/// A pseudo-list of selections (combinations with repetition).
+/// 
+/// A pseudo-list "containing" all the [r]-selections of objects taken from
+/// the list [items].
+/// 
+/// _Example_
+/// 
+///     var s = new Selections(3, characters("abcd"));
+///     print("There are ${s.length} 3-selections of the objects");
+///     print("in ${s.items}.");
+///     print("The first selection is ${c[0]}.");
+/// 
 
 class Selections extends _Combinatorics {
-  int _r;
-
-  /// The number of items taken from [items].
-  int get r => _r;
-
+  
   Selections(int r, List items) {
     if (r < 0)
       throw new Exception("Cannot take $r items from ${items.length}.");
@@ -32,6 +25,11 @@ class Selections extends _Combinatorics {
     _length = _nCr(items.length + r - 1, r);
   }
 
+  int _r;
+
+  /// The number of items taken from [items].
+  int get r => _r;
+
   @override List operator [](int k) =>
       _selection(_adjustedIndex(k, length), r, items);
 
@@ -40,7 +38,8 @@ class Selections extends _Combinatorics {
       contains(selection) ? _inverseSelection(selection, _items) : -1;
 
   /// returns whether [x] is in the pseudo-list.
-  bool contains(List x) => _itemsExistInUniversal(x, _items);
+  @override
+  bool contains(Object x) => _itemsExistInUniversal(x, _items);
 
   @override String toString() =>
       "Pseudo-list containing all $length $r-selections of items from $items.";