+++ lib/src/subsets.dart
@@ -1,20 +1,17 @@
 part of trotter;
 
-/** A pseudolist of subsets.
- * 
- * A pseudo-list "containing" all the subsets of objects taken from
- * the list `elements`.
- * 
- * _Example_
- *
- * ```
- * var sub = new Subsets(characters("abcd"));
- * print("There are ${sub.length} subsets of the objects");
- * print("in ${sub.items}.");
- * print("The first subset is ${sub[0]}.");
- * ```
- * 
- */
+/// A pseudolist of subsets.
+/// 
+/// A pseudo-list "containing" all the subsets of objects taken from
+/// the list `elements`.
+/// 
+/// _Example_
+/// 
+///     var sub = new Subsets(characters("abcd"));
+///     print("There are ${sub.length} subsets of the objects");
+///     print("in ${sub.items}.");
+///     print("The first subset is ${sub[0]}.");
+/// 
 
 class Subsets extends _Combinatorics {
   Subsets(List items) {
@@ -32,7 +29,8 @@ class Subsets extends _Combinatorics {
       contains(subset) ? _inverseSubset(subset, _items) : -1;
 
   /// returns whether [x] is in the pseudo-list.
-  bool contains(List x) =>
+  @override
+  bool contains(Object x) =>
       _itemsExistInUniversal(x, _items) && _itemsAreUnique(x);
 
   @override String toString() =>